refactor(env): extract status rendering/wait flows and keep manager lean
- move status table/panel rendering to environment/status_render.py
- move live/non-terminal wait loop to environment/status_wait.py
- keep EnvironmentMng methods as thin delegates

Fixes: #158
